refactor: merge nlm_service.go into nlm.go and unexport internal types
NLMService, FileChecker, and NewNLMService are only used within
pkg/adapter/nfs/ — unexport them (nlmService, fileChecker, newNLMService)
and consolidate into a single file.
